pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
replit = "^3.2.4"
rasa = "^3.2.5"
logger = "^1.4"
disnake = "^2.5.2"
disnake-jishaku = "^2.6.5"
aiohttp-retry = "^2.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"